bots-discord/python/01-simple-bot/requirements.txt
discord.py>=2.2.0,<3
python-dotenv>=1.*.*
